Remove certtool
No longer needed as most of the functions were unused.
